+++ tests/test_gh/models.py
@@ -1,0 +1,7 @@
+from django.db import models
+
+from localflavor.gh.models import GHRegionField
+
+
+class GHPlace(models.Model):
+    region = GHRegionField(blank=True)