+++ docs/changelog.rst
@@ -11,6 +11,8 @@ New flavors:
 - Belarus localflavor
   (`gh-422 <https://github.com/django/django-localflavor/pull/422>`_,
   `gh-442 <https://github.com/django/django-localflavor/pull/442>`_).
+- Ghana localflavor
+  (`gh-460 <https://github.com/django/django-localflavor/pull/460>`_).
 
 New fields for existing flavors: